move selenium tests to folder
